requirements.txt
apache-airflow==2.5.1
psycopg2-binary==2.9.3
requests==2.28.1
pandas==1.5.2
python-dotenv==0.21.0
boto3==1.26.41
apache-airflow-providers-amazon>=6.0.0